pyproject.toml
[project]
name = "voice-bridge"
version = "0.1.0"
description = "Multi-engine TTS for AI coding assistants — speak responses aloud"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Tomorrow You LLC" }]
keywords = ["tts", "text-to-speech", "claude-code", "ai-assistant", "voice", "accessibility"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Tomorrow-You/voice-bridge"
Repository = "https://github.com/Tomorrow-You/voice-bridge"
Issues = "https://github.com/Tomorrow-You/voice-bridge/issues"

[project.optional-dependencies]
edge = ["edge-tts>=6.1.0"]
elevenlabs = ["elevenlabs>=2.39.0", "sounddevice>=0.5.0", "numpy>=1.26.0"]
kokoro = ["kokoro-onnx>=0.4.0", "sounddevice>=0.5.0", "numpy>=1.26.0"]
all = ["voice-bridge[edge,elevenlabs,kokoro]"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0"]

[project.scripts]
voice-bridge = "voice_bridge.cli:main"
vb-speak = "voice_bridge.speak:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
